Refactor plugin and translations for the new feature
