+++ src/components/TaskInfo/index.js
@@ -2,6 +2,8 @@ import { KeyboardNav } from "../cerebro-ui";
 import { TaskNavItem } from "../index";
 import { completeTask, getTaskHour } from "../../core-engine/taskServices";
 import styles from "./styles.css";
+import lang from "../../lang";
+const strings = lang.TaskInfo;
 
 const TasksInfoPreview = ({ task, actions, client }) => {
 	const action = (task) => {
@@ -17,7 +19,7 @@ const TasksInfoPreview = ({ task, actions, client }) => {
 					<TaskNavItem>{getTaskHour(task) || null}</TaskNavItem>
 					<TaskNavItem>{task.description || null}</TaskNavItem>
 					<TaskNavItem onSelect={() => action(task)}>
-						Completar Tarea
+						{strings.completeTaskButton}
 					</TaskNavItem>
 				</ul>
 			</KeyboardNav>
