+++ src/components/index.js
@@ -1,6 +1,6 @@
-import TaskInfoPreview from "./TasksInfoPreview";
+import TaskInfo from "./TaskInfo";
 import NewTodayTask from "./NewTodayTask";
 import notification from "./notification";
-import TaskNavItem from "./TasksInfoPreview/TaskNavInfoElement";
+import TaskNavItem from "./TaskInfo/TaskNavInfoElement";
 
-export { TaskNavItem, TaskInfoPreview, NewTodayTask, notification };
+export { TaskNavItem, TaskInfo, NewTodayTask, notification };
