+++ src/lang/fr.json
@@ -6,6 +6,7 @@
     "workflow_view": "Todoist: Tâches d'un autre jour",
     "workflow_update": "Mise à jour installée!",
     "invalidCommand": "Commande non valide",
+    "gettingTasksMessage": "Récupération des tâches...",
     
     "settings": {
         "descriptionToken": "Votre clé API Todoist",
@@ -32,16 +33,17 @@
         "li2": "Utilisez :: pour ajouter une description."
     },
 
-    "TodayTasks": {
-        "error": "Une erreur s'est produite (vérifiez votre connexion Internet ou votre clé API Todoist).",
-        "todayTasks" :"Vos tâches aujourd'hui",
-        "noTasks":"Tu n'as pas de tâches aujourd'hui! Est-ce un #TodoistZero ?"
+    "TaskInfo": {
+        "completeTaskButton": "Terminer la tâche",
+        "error": "Une erreur s'est produite (vérifiez votre connexion Internet ou votre clé API Todoist)."
     },
 
-    "XDayTasks": {
-        "error": "Une erreur s'est produite (vérifiez votre connexion Internet ou votre clé API Todoist).",
-        "dayTasks" :"Tâches pour le",
-        "noTasks":"Il n'y a pas de tâches pour ce jour."
+    "displayArrayGenerator":{
+        "noFilteredTasksFound": "Aucune tâche avec ce contenu n'a été trouvée",
+        "noTodayTasks":"Tu n'as pas de tâches aujourd'hui! Est-ce un #TodoistZero ?",
+        "noXDayTasks": "Il n'y a pas de tâches pour ce jour.",
+        "dateNeeded": "Veuillez saisir une date",
+        "invalidDate": "Veuillez entrer une date valide"
     },
 
     "afterUpdate": {
